+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Fixed generals buttons behaviour with the crescendo in add mode https://github.com/GrandOrgue/grandorgue/issues/1209
 - Fixed an empty stop set to a general combination https://github.com/GrandOrgue/grandorgue/issues/1212
 # 3.8.0 (2022-09-15)
 - Fixed setting an empty stop set to a divisional combination https://github.com/GrandOrgue/grandorgue/issues/1068